requirements.txt
faiss-cpu
groq
langchain-groq
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
requests
